INTERLOK-268 AMQP implementation based on QPID INTERLOK-268 #time 2d
Initial checkin of QPID Vendor Implementation; which is a simpler model
than the SwiftMQ AMQP implementation (as in people will understand it
better).

The tests are not as exhaustive as I would like. For some strange reason
if there is a separate broker (activeMQ-5.10) then the AMQP1.0 tests
work, similarly with the AMQP-0.10 tests with a separate QPID broker
running. These are tests with a full send/receive cycle. However as soon
as we try to "embed" the brokers, then the send/receive tests fail.

However because we're only writing a vendorImpl and is only really responsible
for creating the connection factory, we can test without making a
connection. The AMQP1.0 tests attempt to start the connection (it's
creating the session that fails) and check that a javax.jms.Connection
is available; the AMQP-0.10 just check the connection factory.